+++ src/elisa/analytics/binary/mcmc.py
@@ -4,32 +4,31 @@ import os
 import os.path as op
 import json
 
+from abc import ABCMeta, abstractmethod
 from multiprocessing import Pool
 from typing import Iterable, Dict
 from datetime import datetime
 
+from ...conf.config import BINARY_COUNTERPARTS
 from ..binary.plot import Plot
 from ...conf import config
 from ...logger import getPersistentLogger
-from ...base.error import (
-    ElisaError,
-    SolutionBubbleException
-)
+from ...base.error import ElisaError
 from ..binary import (
     utils as analutils,
     params,
-    models,
-    shared
+    models
+)
+from ...analytics.binary.shared import (
+    AbstractLightCurveDataMixin,
+    AbstractCentralRadialVelocityDataMixin,
+    AbstractFit
 )
 
 logger = getPersistentLogger('analytics.binary.mcmc')
 
 
 class McMcMixin(object):
-    @staticmethod
-    def ln_prior(xn):
-        return np.all(np.bitwise_and(np.greater_equal(xn, 0.0), np.less_equal(xn, 1.0)))
-
     @staticmethod
     def resolve_mcmc_result(sampler, labels, discard=False, thin=1, quantiles=None):
         flat_samples = sampler.get_chain(discard=discard, thin=thin, flat=True)
@@ -85,128 +84,123 @@ class McMcMixin(object):
         with open(fpath, "r") as f:
             return json.loads(f.read())
 
+    @staticmethod
+    def worker(sampler, p0, nsteps, nsteps_burn_in):
+        logger.info("running burn-in...")
+        p0, _, _ = sampler.run_mcmc(p0, nsteps_burn_in if nsteps > nsteps_burn_in else nsteps)
+        sampler.reset()
+        logger.info("running production...")
+        _, _, _ = sampler.run_mcmc(p0, nsteps)
 
-class LightCurveFit(shared.AbstractLightCurveFit, McMcMixin):
+
+class McMcFit(AbstractFit, McMcMixin, metaclass=ABCMeta):
     def __init__(self):
         self.plot = Plot()
         self.last_sampler = emcee.EnsembleSampler
         self.last_normalization = dict()
         self.last_fname = ''
-        super(LightCurveFit, self).__init__()
 
-    def model_to_fit(self, *args, **kwargs):
-        return self.likelihood(*args, **kwargs)
+    @staticmethod
+    def ln_prior(xn):
+        return np.all(np.bitwise_and(np.greater_equal(xn, 0.0), np.less_equal(xn, 1.0)))
 
+    @abstractmethod
     def likelihood(self, xn):
-        xn = params.param_renormalizer(xn, self._labels)
-        kwargs = params.prepare_kwargs(xn, self._labels, self._constraint, self._fixed)
-
-        args = self._xs, self._period, self._discretization, self._morphology, self._observer, True
-        synthetic = models.synthetic_binary(*args, **kwargs)
-        synthetic = analutils.normalize_lightcurve_to_max(synthetic)
-
-        lhood = -0.5 * np.sum(np.array([np.sum(np.power((synthetic[band] - self._ys[band]) / self._yerrs[band], 2))
-                                        for band in synthetic]))
-
-        if np.all(np.less_equal(np.abs(lhood), self._xtol)):
-            import sys
-            sys.tracebacklimit = 0
-            raise SolutionBubbleException(f"mcmc hit solution", solution=kwargs)
-
-        return lhood
+        pass
 
     def ln_probability(self, xn):
         if not self.ln_prior(xn):
             return -np.inf
-
         try:
-            likelihood = self.model_to_fit(xn)
-        except SolutionBubbleException:
-            raise
+            likelihood = self.likelihood(xn)
         except (ElisaError, ValueError) as e:
             logger.warning(f'mcmc hit invalid parameters, exception: {str(e)}')
             return -10.0 * np.finfo(float).eps * np.sum(xn)
-
         return likelihood
 
+    def _fit(self, x0, labels, nwalkers, ndim, nsteps, nsteps_burn_in, p0=None):
+
+        p0 = p0 if p0 is not None else np.random.uniform(0.0, 1.0, (nwalkers, ndim))
+        # assign intial value
+        p0[0] = x0
+
+        lnf = self.ln_probability
+        if config.NUMBER_OF_MCMC_PROCESSES > 1:
+            with Pool(processes=config.NUMBER_OF_MCMC_PROCESSES) as pool:
+                logger.info('starting parallel mcmc')
+                sampler = emcee.EnsembleSampler(nwalkers=nwalkers, ndim=ndim, log_prob_fn=lnf, pool=pool)
+                self.worker(sampler, p0, nsteps, nsteps_burn_in)
+        else:
+            logger.info('starting singlecore mcmc')
+            sampler = emcee.EnsembleSampler(nwalkers=nwalkers, ndim=ndim, log_prob_fn=lnf)
+            self.worker(sampler, p0, nsteps, nsteps_burn_in)
+
+        self.last_sampler = sampler
+        self.last_normalization = params.NORMALIZATION_MAP.copy()
+        self.last_fname = self._store_flat_chain(sampler.get_chain(flat=True), labels, self.last_normalization)
+
+        return sampler
+
+
+class LightCurveFit(McMcFit, AbstractLightCurveDataMixin):
+    def likelihood(self, xn):
+        """
+        Liklehood function which defines goodnes of current `xn` vector to be fit of given model.
+        Best is 0.0, worst is -inf.
+
+        :param xn: Iterable[float]; vector of parameters we are looking for
+        :return: float;
+        """
+        xn = params.param_renormalizer(xn, self.labels)
+        kwargs = params.prepare_kwargs(xn, self.labels, self.constraint, self.fixed)
+
+        args = self.xs, self.period, self.discretization, self.morphology, self.observer, True
+        synthetic = models.synthetic_binary(*args, **kwargs)
+        synthetic = analutils.normalize_lightcurve_to_max(synthetic)
+
+        lhood = -0.5 * np.sum(np.array([np.sum(np.power((synthetic[band][self.xs_band_reverser[band]] - self.ys[band])
+                                                        / self.yerrs[band], 2)) for band in synthetic]))
+        return lhood
+
     def fit(self, xs, ys, period, x0, discretization, nwalkers, nsteps,
-            xtol=1e-6, p0=None, yerrs=None, nsteps_burn_in=10, quantiles=None, discard=False):
+            p0=None, yerrs=None, nsteps_burn_in=10, quantiles=None, discard=False):
         """
         Fit method using Markov Chain Monte Carlo.
+        Once simulation is done, following valeus are stored and can be used for further evaluation::
 
-        :param xs: Iterable[float];
-        :param ys: Dict;
+            self.last_sampler: emcee.EnsembleSampler
+            self.last_normalization: Dict; normalization map used during fitting
+            self.last_fname: str; filename of last stored flatten emcee `sampler` with metadata
+
+        :param xs: Dict[str, Iterable[float]]; {<passband>: <phases>}
+        :param ys: Dict[str, Iterable[float]]; {<passband>: <fluxes>};
         :param period: float; sytem period
         :param x0: List[Dict]; initial state (metadata included)
         :param discretization: float; discretization of objects
         :param nwalkers: int; number of walkers
         :param nsteps: int; number of steps in mcmc eval
-        :param xtol: float; tolerance of error to consider hitted solution as exact
         :param p0: numpy.array; inital priors for mcmc
         :param yerrs: Union[numpy.array, float]; errors for each point of observation
         :param nsteps_burn_in: int; numer of steps for mcmc to explore parameters
         :param quantiles: List[int];
         :param discard: Union[int, bool]; how many values of result discard when looking for solution
-        :return: Dict; solution on supplied quantiles, default is [16, 50, 84]
+        :return: emcee.EnsembleSampler; sampler instance
         """
 
-        def worker(_sampler, _p0):
-            logger.info("running burn-in...")
-            _p0, _, _ = sampler.run_mcmc(_p0, nsteps_burn_in if nsteps > nsteps_burn_in else nsteps)
-            _sampler.reset()
-            logger.info("running production...")
-            _, _, _ = sampler.run_mcmc(_p0, nsteps)
-
-        passband = list(ys.keys())
-        yerrs = {band: analutils.lightcurves_mean_error(ys) for band in passband} if yerrs is None else yerrs
-        # xs = xs if isinstance(xs, dict) else {band: xs for band in ys}
-        self._xs, self._ys, self._yerrs = xs, ys, yerrs
-        self._xtol = xtol
-
-        x0 = params.initial_x0_validity_check(x0, self._morphology)
-        x0, labels, fixed, constraint, observer = params.fit_data_initializer(x0, passband=passband)
+        self.passband = list(ys.keys())
+        yerrs = {band: analutils.lightcurves_mean_error(ys) for band in self.passband} if yerrs is None else yerrs
+        x0 = params.initial_x0_validity_check(x0, self.morphology)
+        x0, labels, fixed, constraint, observer = params.fit_data_initializer(x0, passband=self.passband)
         ndim = len(x0)
+        params.mcmc_nwalkers_vs_ndim_validity_check(nwalkers, ndim)
 
-        if nwalkers < ndim * 2:
-            msg = f'Fit cannot be executed with fewer walkers ({nwalkers}) than twice the number of dimensions ({ndim})'
-            raise RuntimeError(msg)
+        self.xs, self.xs_band_reverser = params.xs_reducer(xs)
+        self.labels, self.observer, self.period = labels, observer, period
+        self.fixed, self.constraint = fixed, constraint
+        self.ys, self.yerrs = ys, yerrs
+        self.discretization = discretization
 
-        self._hash_map = {key: idx for idx, key in enumerate(labels)}
-        self._period = period
-        self._morphology = self._morphology
-        self._discretization = discretization
-        self._passband = passband
-        self._fixed = fixed
-        self._constraint = constraint
-        self._labels = labels
-        self._observer = observer
-
-        p0 = p0 if p0 is not None else np.random.uniform(0.0, 1.0, (nwalkers, ndim))
-        # assign intial value
-        p0[0] = x0
-        try:
-            lnf = self.ln_probability
-            if config.NUMBER_OF_MCMC_PROCESSES > 1:
-                with Pool(processes=config.NUMBER_OF_MCMC_PROCESSES) as pool:
-                    logger.info('starting parallel mcmc')
-                    sampler = emcee.EnsembleSampler(nwalkers=nwalkers, ndim=ndim, log_prob_fn=lnf, pool=pool)
-                    worker(sampler, p0)
-            else:
-                logger.info('starting singlecore mcmc')
-                sampler = emcee.EnsembleSampler(nwalkers=nwalkers, ndim=ndim, log_prob_fn=lnf)
-                worker(sampler, p0)
-
-        except SolutionBubbleException as bubble:
-            result = self.serialize_bubble(bubble)
-            return params.extend_result_with_units(result)
-
-        result = self.resolve_mcmc_result(sampler, labels, discard=discard)
-        result = result + [{"param": key, "value": val} for key, val in self._fixed.items()]
-
-        self.last_sampler = sampler
-        self.last_normalization = params.NORMALIZATION_MAP.copy()
-        self.last_fname = self._store_flat_chain(sampler.get_chain(flat=True), labels, self.last_normalization)
-        return params.extend_result_with_units(result)
+        return self._fit(x0, self.labels, nwalkers, ndim, nsteps, nsteps_burn_in, p0)
 
 
 class OvercontactLightCurveFit(LightCurveFit):
@@ -216,7 +210,7 @@ class OvercontactLightCurveFit(LightCurveFit):
     """
     def __init__(self):
         super().__init__()
-        self._morphology = 'over-contact'
+        self.morphology = 'over-contact'
 
 
 class DetachedLightCurveFit(LightCurveFit):
@@ -225,25 +219,62 @@ class DetachedLightCurveFit(LightCurveFit):
     """
     def __init__(self):
         super().__init__()
-        self._morphology = 'detached'
+        self.morphology = 'detached'
 
 
-class CentralRadialVelocity(shared.AbstractCentralRadialVelocity, McMcMixin):
-    def __init__(self):
-        self.plot = Plot()
-        self.last_sampler = emcee.EnsembleSampler
-        self.last_normalization = dict()
-        self.last_fname = ''
-        super(CentralRadialVelocity, self).__init__()
-
+class CentralRadialVelocity(McMcFit, AbstractCentralRadialVelocityDataMixin):
     def likelihood(self, xn):
-        pass
+        """
+        Liklehood function which defines goodnes of current `xn` vector to be fit of given model.
+        Best is 0.0, worst is -inf.
 
-    def ln_probability(self, xn):
-        pass
+        :param xn: Iterable[float]; vector of parameters we are looking for
+        :return: float;
+        """
+        xn = params.param_renormalizer(xn, self.labels)
+        kwargs = params.prepare_kwargs(xn, self.labels, self.constraint, self.fixed)
 
-    def fit(self):
-        pass
+        args = self.xs, self.period, self.observer
+        synthetic = models.central_rv_synthetic(*args, **kwargs)
+        if self.on_normalized:
+            synthetic = analutils.normalize_rv_curve_to_max(synthetic)
+        synthetic = {"primary": synthetic[0], "secondary": synthetic[1]}
+
+        lhood = -0.5 * np.sum(np.array([np.sum(np.power((synthetic[comp] - self.ys[comp]) / self.yerrs[comp], 2))
+                                        for comp in BINARY_COUNTERPARTS]))
+        return lhood
+
+    def fit(self, xs, ys, period, x0, nwalkers, nsteps, p0=None, yerrs=None, nsteps_burn_in=10):
+        """
+        Fit method using Markov Chain Monte Carlo.
+        Once simulation is done, following valeus are stored and can be used for further evaluation::
+
+            self.last_sampler: emcee.EnsembleSampler
+            self.last_normalization: Dict; normalization map used during fitting
+            self.last_fname: str; filename of last stored flatten emcee `sampler` with metadata
+
+        :param xs: Iterable[float];
+        :param ys: Dict;
+        :param period: float; sytem period
+        :param x0: List[Dict]; initial state (metadata included)
+        :param nwalkers: int; number of walkers
+        :param nsteps: int; number of steps in mcmc eval
+        :param p0: numpy.array; inital priors for mcmc
+        :param yerrs: Union[numpy.array, float]; errors for each point of observation
+        :param nsteps_burn_in: int; numer of steps for mcmc to explore parameters
+        :return: emcee.EnsembleSampler; sampler instance
+        """
+
+        yerrs = {c: analutils.radialcurves_mean_error(ys) for c in BINARY_COUNTERPARTS} if yerrs is None else yerrs
+        x0, labels, fixed, constraint, observer = params.fit_data_initializer(x0)
+        ndim = len(x0)
+        params.mcmc_nwalkers_vs_ndim_validity_check(nwalkers, ndim)
+
+        self.labels, self.observer, self.period = labels, observer, period
+        self.fixed, self.constraint = fixed, constraint
+        self.xs, self.ys, self.yerrs = xs, ys, yerrs
+
+        return self._fit(x0, self.labels, nwalkers, ndim, nsteps, nsteps_burn_in, p0)
 
 
 binary_detached = DetachedLightCurveFit()