+++ unittests/test_analytics.py
@@ -1,21 +1,22 @@
 import time
-
-import numpy as np
-
 from copy import copy
-from numpy.testing import assert_array_equal
-from elisa.analytics.binary import params
-from elisa.base.error import InitialParamsError
-from unittests.utils import ElisaTestCase
-
 from unittest import (
     mock
 )
+
+import numpy as np
+from numpy.testing import assert_array_equal
+
+from elisa.analytics.binary import params
 from elisa.analytics.binary.least_squares import (
     binary_detached as ls_binary_detached,
     central_rv
 )
 from elisa.analytics.binary.mcmc import binary_detached as mc_binary_detached
+from elisa.analytics.binary.mcmc import central_rv as mc_central_rv
+from elisa.base.error import InitialParamsError
+from elisa.conf.config import BINARY_COUNTERPARTS
+from unittests.utils import ElisaTestCase
 
 TOL = 1e-5
 
@@ -120,25 +121,18 @@ class TestParamsTestCase(ElisaTestCase):
         xn = [
             {
                 "value": 1.1,
-                "param": "p__surface_potential",
-                "min": 1.0,
-                "max": 10.0
+                "param": "p__surface_potential"
             },
             {
                 "value": 1.2,
-                "param": "s__surface_potential",
-                "min": 1.1,
-                "max": 11.0
+                "param": "s__surface_potential"
             }
         ]
 
-        expected = [{'value': 1.1, 'param': 'p__surface_potential', 'min': 1.0, 'max': 10.0},
-                    {'param': 's__surface_potential', 'value': 1.1, 'min': 1.0, 'max': 10.0}]
+        expected = [{'value': 1.1, 'param': 'p__surface_potential'},
+                    {'param': 's__surface_potential', 'value': 1.1}]
         obtained = params.adjust_result_constrained_potential(xn, hash_map)
-
         self.assertTrue(expected[0]['value'] == obtained[0]['value'])
-        self.assertTrue(expected[0]['min'] == obtained[0]['min'])
-        self.assertTrue(expected[0]['max'] == obtained[0]['max'])
 
     def test_initial_x0_validity_check(self):
         xn = [
@@ -225,15 +219,38 @@ class TestParamsTestCase(ElisaTestCase):
         vectorized, labels = params.x0_vectorize(x0)
         x0v = {key: val for key, val in zip(labels, vectorized)}
 
-        evaluated = params.eval_constraints(x0v, x0c)
+        evaluated = params.constraints_evaluator(x0v, x0c)
         self.assertTrue(evaluated["s__mass"] == 20)
+
+    def test_xs_reducer_all_same(self):
+        a = np.linspace(0, 1, 5)
+        xs = {"a": a, "b": a, "c": a}
+        new_xs, inverse = params.xs_reducer(xs)
+        for band, phases in xs.items():
+            assert_array_equal(inverse[band], np.arange(len(phases)))
+        assert_array_equal(new_xs, a)
+
+    def test_xs_reducer_random(self):
+        a = np.linspace(0, 1, 5)
+        b = np.array(np.array(a + 0.01).tolist() + [0.021])
+        c = a + 0.02
+        xs = {"a": a, "b": b, "c": c}
+        new_xs, inverse = params.xs_reducer(xs)
+        expected_xs = [0., 0.01, 0.02, 0.021, 0.25, 0.26, 0.27, 0.5, 0.51, 0.52, 0.75, 0.76, 0.77, 1., 1.01, 1.02]
+        expected_inverse = {'a': [0, 4, 7, 10, 13], 'b': [1, 5, 8, 11, 14, 3], 'c': [2, 6, 9, 12, 15]}
+
+        for band, phases in xs.items():
+            assert_array_equal(inverse[band], expected_inverse[band])
+        assert_array_equal(expected_xs, new_xs)
 
 
 class AbstractFitTestCase(ElisaTestCase):
     def setUp(self):
         self.model_generator = ModelSimulator()
 
-    phases = np.arange(-0.6, 0.62, 0.02)
+    phases = {'Generic.Bessell.V': np.arange(-0.6, 0.62, 0.02),
+              'Generic.Bessell.B': np.arange(-0.6, 0.62, 0.02)}
+
     flux = {'Generic.Bessell.V': np.array([0.98128349, 0.97901564, 0.9776404, 0.77030991, 0.38623294,
                                            0.32588823, 0.38623294, 0.77030991, 0.9776404, 0.97901564,
                                            0.98128349, 0.9831816, 0.98542223, 0.9880625, 0.99034951,
@@ -266,7 +283,6 @@ class McMcLCTestCase(AbstractFitTestCase):
     """
     Requre just methods to pass.
     """
-
     def test_mcmc_lc_fit_std_params_detached(self):
         dinit = [
             {
@@ -311,9 +327,9 @@ class McMcLCTestCase(AbstractFitTestCase):
         ]
 
         self.model_generator.keep_out = True
-        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.generator):
+        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.lc_generator):
             mc_binary_detached.fit(xs=self.phases, ys=self.flux, period=3.0, discretization=5,
-                                   x0=dinit, nwalkers=4, nsteps=5, xtol=1e-5)
+                                   x0=dinit, nwalkers=4, nsteps=5)
 
     def test_mcmc_lc_fit_community_params_detached(self):
         dinit = [
@@ -363,12 +379,12 @@ class McMcLCTestCase(AbstractFitTestCase):
             }
         ]
         self.model_generator.keep_out = True
-        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.generator):
+        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.lc_generator):
             mc_binary_detached.fit(xs=self.phases, ys=self.flux, period=3.0, discretization=5,
-                                   x0=dinit, nwalkers=4, nsteps=5, xtol=1e-5)
+                                   x0=dinit, nwalkers=4, nsteps=5)
 
 
-class LeastSqaureRVTestCase(ElisaTestCase):
+class RVTestCase(ElisaTestCase):
     rv = {'primary': np.array([111221.02018955, 102589.40515112, 92675.34114568,
                                81521.98280508, 69189.28515476, 55758.52165462,
                                41337.34984718, 26065.23187763, 10118.86370365,
@@ -412,6 +428,53 @@ class LeastSqaureRVTestCase(ElisaTestCase):
                                  9082.58262586, 37786.04533903, 67309.27849613,
                                  97176.13649135])}
 
+
+class McMcRVTestCase(RVTestCase):
+    def test_mcmc_rv_fit_community_params(self):
+        phases = np.arange(-0.6, 0.62, 0.02)
+        initial_parameters = [
+            {
+                'value': 0.1,
+                'param': 'eccentricity',
+                'fixed': True,
+
+            },
+            {
+                'value': 20.0,  # 4.219470628180749
+                'param': 'asini',
+                'fixed': False,
+                'min': 1.0,
+                'max': 100
+
+            },
+            {
+                'value': 0.8,  # 1.0 / 1.8
+                'param': 'mass_ratio',
+                'fixed': False,
+                'min': 0,
+                'max': 2
+            },
+            {
+                'value': 0.0,
+                'param': 'argument_of_periastron',
+                'fixed': True
+            },
+            {
+                'value': 20000.0,
+                'param': 'gamma',
+                'fixed': True
+            }
+        ]
+
+        model_generator = ModelSimulator()
+        model_generator.keep_out = True
+        with mock.patch("elisa.analytics.binary.models.central_rv_synthetic", model_generator.rv_generator):
+            mc_central_rv.fit(xs=phases, ys=self.rv, period=0.6, x0=initial_parameters, nwalkers=4, nsteps=5)
+        # result = central_rv.fit(xs=phases, ys=self.rv, period=0.6, x0=copy(initial_parameters))
+        # self.assertTrue(1.0 > result[-1]["r_squared"] > 0.9)
+
+
+class LeastSqaureRVTestCase(RVTestCase):
     def test_least_squares_rv_fit_std_params(self):
         """
         Test has to pass and finis in real time.
@@ -544,7 +607,7 @@ class LeastSqaureLCTestCase(AbstractFitTestCase):
             }
         ]
 
-        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.generator):
+        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.lc_generator):
             result = ls_binary_detached.fit(self.phases, self.flux, period=3.0, discretization=5, x0=dinit, xtol=1e-5)
         self.assertTrue(1.0 > result[-1]["r_squared"] > 0.9)
 
@@ -595,7 +658,7 @@ class LeastSqaureLCTestCase(AbstractFitTestCase):
                 'fixed': True
             }
         ]
-        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.generator):
+        with mock.patch("elisa.analytics.binary.models.synthetic_binary", self.model_generator.lc_generator):
             result = ls_binary_detached.fit(self.phases, self.flux, period=3.0, discretization=5, x0=dinit, xtol=1e-5)
         self.assertTrue(1.0 > result[-1]["r_squared"] > 0.9)
 
@@ -627,7 +690,52 @@ class ModelSimulator(object):
                                            0.80924345, 0.80729325, 0.80604709, 0.60603475, 0.2294959,
                                            0.17384023, 0.2294959, 0.60603475, 0.80604709, 0.80729325,
                                            0.80924345])}
-    mean = np.mean(list(flux.values()))
+
+    rv = {'primary': np.array([111221.02018955, 102589.40515112, 92675.34114568,
+                               81521.98280508, 69189.28515476, 55758.52165462,
+                               41337.34984718, 26065.23187763, 10118.86370365,
+                               -6282.93249474, -22875.63138097, -39347.75579673,
+                               -55343.14273712, -70467.93174445, -84303.24303593,
+                               -96423.915992, -106422.70195531, -113938.05716509,
+                               -118682.43573797, -120467.13803823, -119219.71866465,
+                               -114990.89801808, -107949.71016039, -98367.77975255,
+                               -86595.51823899, -73034.14124119, -58107.52819161,
+                               -42237.21613808, -25822.61388457, -9227.25025377,
+                               7229.16722243, 23273.77242388, 38679.82691287,
+                               53263.47669152, 66879.02978007, 79413.57620399,
+                               90781.53548261, 100919.51001721, 109781.66096297,
+                               117335.70723602, 123559.57210929, 128438.6567666,
+                               131963.69775175, 134129.15836278, 134932.10727626,
+                               134371.54717101, 132448.1692224, 129164.5242993,
+                               124525.61727603, 118539.94602416, 111221.02018955,
+                               102589.40515112, 92675.34114568, 81521.98280508,
+                               69189.28515476, 55758.52165462, 41337.34984718,
+                               26065.23187763, 10118.86370365, -6282.93249474,
+                               -22875.63138097]),
+          'secondary': np.array([-144197.83633559, -128660.92926642, -110815.61405663,
+                                 -90739.56904355, -68540.71327298, -44365.33897272,
+                                 -18407.22971932, 9082.58262586, 37786.04533903,
+                                 67309.27849613, 97176.13649135, 126825.96043971,
+                                 155617.65693242, 182842.27714561, 207745.83747028,
+                                 229563.04879121, 247560.86352515, 261088.50290277,
+                                 269628.38433395, 272840.84847442, 270595.49360196,
+                                 262983.61643814, 250309.4782943, 233062.00356019,
+                                 211871.93283578, 187461.45423975, 160593.55075051,
+                                 132026.98905414, 102480.70499782, 72609.05046238,
+                                 42987.49900522, 14107.20964261, -13623.68843756,
+                                 -39874.25803914, -64382.25359853, -86944.43716158,
+                                 -107406.76386309, -125655.11802538, -141606.98972774,
+                                 -155204.27301923, -166407.22979112, -175189.58217428,
+                                 -181534.65594755, -185432.4850474, -186877.79309168,
+                                 -185868.78490222, -182406.70459473, -176496.14373313,
+                                 -168146.11109126, -157371.90283788, -144197.83633559,
+                                 -128660.92926641, -110815.61405663, -90739.56904355,
+                                 -68540.71327297, -44365.33897271, -18407.22971932,
+                                 9082.58262586, 37786.04533903, 67309.27849613,
+                                 97176.13649135])}
+
+    lc_mean = np.mean(list(flux.values()))
+    rv_mean = np.mean(list(rv.values()))
 
     def __init__(self, keep_out=False):
         self.error = 0.1
@@ -635,10 +743,10 @@ class ModelSimulator(object):
         self.args = []
         self.keep_out = keep_out
 
-    def generator(self, *args, **kwargs):
+    def _prepare(self, *args, **kwargs):
         np.random.seed(1)
         self.args.append({"args": args, "kwargs": kwargs})
-        self.error = np.exp(-(self.step * 0.25)**2)
+        self.error = np.exp(-(self.step * 0.25) ** 2)
 
         if self.keep_out:
             np.random.seed(int(time.time()))
@@ -646,7 +754,100 @@ class ModelSimulator(object):
 
         self.step += 1
 
-        add = self.mean * self.error
+    def lc_generator(self, *args, **kwargs):
+        self._prepare(*args, **kwargs)
+        add = self.lc_mean * self.error
         flux = {band: self.flux[band] + np.random.uniform(-add, +add, len(self.flux[band]))
                 for band in self.flux}
         return flux
+
+    def rv_generator(self, *args, **kwargs):
+        self._prepare(*args, **kwargs)
+        add = self.rv_mean * self.error
+        rv = {component: self.rv[component] + np.random.uniform(-add, +add, len(self.rv[component]))
+              for component in BINARY_COUNTERPARTS}
+        return rv["primary"], rv["secondary"]
+
+
+class ConstraintsTestCase(ElisaTestCase):
+    basic_x0 = [
+        {
+            "param": "a",
+            "value": 50
+        },
+        {
+            "param": "b",
+            "value": 2
+        },
+        {
+            "param": "c",
+            "value": 1.25
+        },
+        {
+            "param": "c1",
+            "value": 0,
+            "constraint": "2 * {a}"
+        },
+        {
+            "param": "c2",
+            "value": 0,
+            "constraint": "2 + {b}"
+        },
+        {
+            "param": "c3",
+            "value": 0,
+            "constraint": "{a} + {b}"
+        },
+        {
+            "param": "c4",
+            "value": 0,
+            "constraint": "({a} + {b}) * 2.0"
+        }
+    ]
+
+    extended_x0 = [
+        {
+            "param": "a",
+            "value": 50
+        },
+        {
+            "param": "b",
+            "value": 2
+        },
+        {
+            "param": "c1",
+            "value": 0,
+            "constraint": "{a} * sin({b})"
+        },
+        {
+            "param": "c2",
+            "value": 0,
+            "constraint": "{a} * log({b})"
+        },
+        {
+            "param": "c2",
+            "value": 0,
+            "constraint": "{a} * log10({b})"
+        }
+    ]
+
+    def test_constraints_validator_basic(self):
+        params.constraints_validator(self.basic_x0)
+
+    def test_constraints_evaluator_basic(self):
+        expected = {'c1': 100, 'c2': 4, 'c3': 52, 'c4': 104.0}
+        floats = params.x0_to_floats_kwargs(self.basic_x0)
+        constraints = params.x0_to_constraints_kwargs(self.basic_x0)
+        obtained = params.constraints_evaluator(floats, constraints)
+        self.assertDictEqual(expected, obtained)
+
+    def test_constraints_validator_extended(self):
+        params.constraints_validator(self.extended_x0)
+
+    def test_constraints_evaluator_extended(self):
+        expected = {'c1': np.round(45.46487134128409, 3), 'c2': np.round(15.05149978319906, 3)}
+        floats = params.x0_to_floats_kwargs(self.extended_x0)
+        constraints = params.x0_to_constraints_kwargs(self.extended_x0)
+        obtained = params.constraints_evaluator(floats, constraints)
+        obtained = {key: np.round(val, 3) for key, val in obtained.items()}
+        self.assertDictEqual(expected, obtained)